+++ lib/output/writer/elasticsearch.go
@@ -28,10 +28,13 @@ import (
 	"time"
 
 	"github.com/Jeffail/benthos/lib/log"
+	"github.com/Jeffail/benthos/lib/message"
 	"github.com/Jeffail/benthos/lib/metrics"
 	"github.com/Jeffail/benthos/lib/types"
 	"github.com/Jeffail/benthos/lib/util/http/auth"
+	"github.com/Jeffail/benthos/lib/util/retries"
 	"github.com/Jeffail/benthos/lib/util/text"
+	"github.com/cenkalti/backoff"
 	"github.com/olivere/elastic"
 )
 
@@ -40,18 +43,24 @@ import (
 // ElasticsearchConfig contains configuration fields for the Elasticsearch
 // output type.
 type ElasticsearchConfig struct {
-	URLs      []string             `json:"urls" yaml:"urls"`
-	Sniff     bool                 `json:"sniff" yaml:"sniff"`
-	ID        string               `json:"id" yaml:"id"`
-	Index     string               `json:"index" yaml:"index"`
-	Pipeline  string               `json:"pipeline" yaml:"pipeline"`
-	Type      string               `json:"type" yaml:"type"`
-	TimeoutMS int                  `json:"timeout_ms" yaml:"timeout_ms"`
-	Auth      auth.BasicAuthConfig `json:"basic_auth" yaml:"basic_auth"`
+	URLs           []string             `json:"urls" yaml:"urls"`
+	Sniff          bool                 `json:"sniff" yaml:"sniff"`
+	ID             string               `json:"id" yaml:"id"`
+	Index          string               `json:"index" yaml:"index"`
+	Pipeline       string               `json:"pipeline" yaml:"pipeline"`
+	Type           string               `json:"type" yaml:"type"`
+	TimeoutMS      int                  `json:"timeout_ms" yaml:"timeout_ms"`
+	Auth           auth.BasicAuthConfig `json:"basic_auth" yaml:"basic_auth"`
+	retries.Config `json:",inline" yaml:",inline"`
 }
 
 // NewElasticsearchConfig creates a new ElasticsearchConfig with default values.
 func NewElasticsearchConfig() ElasticsearchConfig {
+	rConf := retries.NewConfig()
+	rConf.Backoff.InitialInterval = "1s"
+	rConf.Backoff.MaxInterval = "5s"
+	rConf.Backoff.MaxElapsedTime = "30s"
+
 	return ElasticsearchConfig{
 		URLs:      []string{"http://localhost:9200"},
 		Sniff:     true,
@@ -61,6 +70,7 @@ func NewElasticsearchConfig() ElasticsearchConfig {
 		Type:      "doc",
 		TimeoutMS: 5000,
 		Auth:      auth.NewBasicAuthConfig(),
+		Config:    rConf,
 	}
 }
 
@@ -75,10 +85,14 @@ type Elasticsearch struct {
 	sniff bool
 	conf  ElasticsearchConfig
 
+	backoff backoff.BackOff
+
 	idStr             *text.InterpolatedString
 	indexStr          *text.InterpolatedString
 	pipelineStr       *text.InterpolatedString
 	interpolatedIndex bool
+
+	eJSONErr metrics.StatCounter
 
 	client *elastic.Client
 }
@@ -94,6 +108,7 @@ func NewElasticsearch(conf ElasticsearchConfig, log log.Modular, stats metrics.T
 		indexStr:          text.NewInterpolatedString(conf.Index),
 		pipelineStr:       text.NewInterpolatedString(conf.Pipeline),
 		interpolatedIndex: text.ContainsFunctionVariables([]byte(conf.Index)),
+		eJSONErr:          stats.GetCounter("output.elasticsearch.error.json"),
 	}
 
 	for _, u := range conf.URLs {
@@ -102,6 +117,11 @@ func NewElasticsearch(conf ElasticsearchConfig, log log.Modular, stats metrics.T
 				e.urls = append(e.urls, splitURL)
 			}
 		}
+	}
+
+	var err error
+	if e.backoff, err = conf.Config.Get(); err != nil {
+		return nil, err
 	}
 
 	return &e, nil
@@ -148,6 +168,13 @@ func (e *Elasticsearch) Connect() error {
 	return err
 }
 
+func shouldRetry(s int) bool {
+	if s >= 500 && s <= 599 {
+		return true
+	}
+	return false
+}
+
 // Write will attempt to write a message to Elasticsearch, wait for
 // acknowledgement, and returns an error if applicable.
 func (e *Elasticsearch) Write(msg types.Message) error {
@@ -155,17 +182,72 @@ func (e *Elasticsearch) Write(msg types.Message) error {
 		return types.ErrNotConnected
 	}
 
-	return msg.Iter(func(i int, part types.Part) error {
+	if msg.Len() == 1 {
+		index := e.indexStr.Get(msg)
 		_, err := e.client.Index().
-			Index(e.indexStr.Get(msg)).
+			Index(index).
 			Pipeline(e.pipelineStr.Get(msg)).
 			Type(e.conf.Type).
 			Id(e.idStr.Get(msg)).
-			BodyString(string(part.Get())).
+			BodyString(string(msg.Get(0).Get())).
 			Do(context.Background())
-
+		if err == nil {
+			// Flush to make sure the document got written.
+			_, err = e.client.Flush().Index(index).Do(context.Background())
+		}
 		return err
+	}
+
+	e.backoff.Reset()
+
+	b := e.client.Bulk()
+	docs := []interface{}{}
+	msg.Iter(func(i int, part types.Part) error {
+		jObj, ierr := part.JSON()
+		if ierr != nil {
+			e.eJSONErr.Incr(1)
+			e.log.Errorf("Failed to marshal message into JSON document: %v\n", ierr)
+			return nil
+		}
+		docs = append(docs, jObj)
+		b.Add(
+			elastic.NewBulkIndexRequest().
+				Index(e.indexStr.Get(message.Lock(msg, i))).
+				Pipeline(e.pipelineStr.Get(message.Lock(msg, i))).
+				Type(e.conf.Type).
+				Id(e.idStr.Get(message.Lock(msg, i))).
+				Doc(jObj),
+		)
+		return nil
 	})
+
+	for b.NumberOfActions() != 0 {
+		wait := e.backoff.NextBackOff()
+
+		result, err := b.Do(context.Background())
+		if err != nil {
+			return err
+		}
+
+		failed := result.Failed()
+		if len(failed) == 0 {
+			return nil
+		}
+
+		for i := 0; i < len(failed); i++ {
+			if !shouldRetry(failed[i].Status) {
+				e.log.Errorf("elasticsearch message rejected with code [%s]: %v\n", failed[i].Status, failed[i].Error.Reason)
+				return fmt.Errorf("failed to send %v parts from message: %v", len(failed), failed[0].Error.Reason)
+			}
+			e.log.Errorf("elasticsearch message failed with code [%s]: %v\n", failed[i].Status, failed[i].Error.Reason)
+		}
+		if wait == backoff.Stop {
+			return fmt.Errorf("failed to send %v parts from message: %v", len(failed), failed[0].Error.Reason)
+		}
+		time.Sleep(wait)
+	}
+
+	return nil
 }
 
 // CloseAsync shuts down the Elasticsearch writer and stops processing messages.